backend/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-filter==22.1
django-tinymce==3.5.0
djangorestframework==3.14.0
drf_extra_fields==3.4.1
drf-yasg==1.21.4
gunicorn==20.0.4
Pillow==9.2.0
pyyaml==6.0
pytz==2022.4
sorl-thumbnail==12.9.0
sqlparse==0.4.3
